requirements.txt
google-generativeai
google-api-python-client
python-dotenv
flask
gunicorn